fix(tests): Make first basic block test agree with test plan
The test plan specified to use a block from the Basic Blocks
category of the test blocks.

PR #7324 changed the test in a way that violated the plan,
albeit in aid of a good cause: the basic blocks category had
become unusable due to a bug in the test blocks combined with
better error checking (in PR #7289).

Since the test blocks were fixed by google/blockly-samples#1774
(and the fix published and Blockly updated by PR #7313) this
violation can be corrected.